+++ pyproject.toml
@@ -1,0 +1,22 @@
+[tool.poetry]
+name = "dotsy"
+version = "0.1.0"
+description = "A GUI tool for managing Linux config files"
+authors = ["Aamir Azad <82281117+aamirazad@users.noreply.github.com>"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = ">=3.9,<3.14"
+GitPython = "^3.1.40"
+PyYAML = "^6.0.1"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^7.4.3"
+black = "^23.11.0"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+dotsy = "dotsy.main:main"
