+++ src/dotsy/main.py
@@ -1,0 +1,30 @@
+# For command line arguments
+import sys
+from PySide6.QtWidgets import QApplication, QLabel, QMainWindow
+
+class MainWindow(QMainWindow):
+    def __init__(self):
+        super().__init__()
+        self.label = QLabel("Click in this window")
+        self.setCentralWidget(self.label)
+
+    def mouseMoveEvent(self, e):
+        self.label.setText("mouseMoveEvent")
+
+    def mousePressEvent(self, e):
+        self.label.setText("mousePressEvent")
+
+    def mouseReleaseEvent(self, e):
+        self.label.setText("mouseReleaseEvent")
+
+    def mouseDoubleClickEvent(self, e):
+        self.label.setText("mouseDoubleClickEvent")
+
+app = QApplication(sys.argv)
+
+window = MainWindow()
+window.show() # Windows are hidden by default
+
+# Start the exec loop
+app.exec()
+# Code after this will only run after exit
